EduDesk/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
plotly==5.15.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
Werkzeug==2.3.7
